+++ CHANGES.rst
@@ -2,13 +2,12 @@
 Changelog
 ========================
 
-0.7.0 - dev
+0.8.0 - dev
 =================
 
 New Features
 ------------
-- widget to save SMLM data in selected file formats
-- widget to show information on current SMLM dataset
+-
 
 API Changes
 -----------
@@ -16,15 +15,27 @@ API Changes
 
 Bug Fixes
 ---------
-- fix documentation
+-
 
 Other Changes and Additions
 ---------------------------
 -
 
+0.7.0 - 2025-03-17
+==================
+
+New Features
+------------
+- widget to save SMLM data in selected file formats
+- widget to show information on current SMLM dataset
+
+Bug Fixes
+---------
+- fix documentation
+
 
 0.6.0 - 2025-03-11
-=================
+==================
 
 New Features
 ------------